Selector Class An S4 class providing the methods to select a model from a list.

The object consists of 3 slots

Constructor for the S4 Selector object.

Usage

Selector(id, selector, logger = Logger(verbose = FALSE))

Arguments

id

object identifier, must be equivalent to the id of the associated Trainer. If learning method is one of the supported by renoir, the constructor will automatically select a built-in selector function.

selector

(optional) function used to select a model from a list. Used if id is not one of the supported by renoir. If selector is provided it must conform to the renoir common interface, and must have the following formal arguments:

models

list of models

...

additional arguments

It should return the index of the selected model

logger

a Logger

Value

An object of class Selector.

Slots

id

the object identifier, it must be equivalent to the learning method name (i.e. the trainer id)

selector

function to use for selecting the model

logger

an object of class Logger

Author

Alessandro Barberis