Create an interactive report

Create an interactive report for the computed analysis.

Usage

# S4 method for Trained
create_report(
  output_format = "html_document",
  outdir = tempdir(),
  filename = NULL,
  report.type = c("full", "short"),
  object,
  x,
  y,
  weights = NULL,
  offset = NULL,
  annotation = NULL,
  feat.index = 0,
  ...
)

Arguments

outdir

path to the output directory

filename

(optional) name without extension for the output file

report.type

the type of report to create. If short, the feature importance tables are not created (this could solve the pandoc "out of memory" issue when the number of considered features is high)

object

an object of class Trained

annotation

a data frame containing features annotation

feat.index

numeric, the index of the column in annotation matching x feature names. 0 means rownames

...

further arguments to render params argument. Possible options are:

yml_author

A character vector, name of the author(s)

yml_affiliation

The author’s affiliation; must match length of name, e.g. if name has length of two, affiliation must as well; use NA if you don’t want to include an affiliation for a given author.Note that not all formats support the affiliation field.

yml_email

The author email address. Note that not all formats support the email field.

yml_title

A character vector, the title of the document

yml_subtitle

A character vector, the subtitle of the document. Not all R Markdown formats use subtitles, so it may depend on what you use in the output field. It is available in pdf_document, html_document, and word_document by default

yml_abstract

A character vector, the abstract. Long character vectors are automatically wrapped using valid YAML syntax. This field is not available in all output formats; it is available in pdf_document and html_document by default.

yml_name

A character vector, the name of the document

tabset

logical, whether to use tabbed sections (if any)

Value

The compiled report is written into the output file.

Author

Alessandro Barberis