Compute the significance of the filtering test

This function compute the significance of the screening test by using an Empirical Bayes method.

Usage

ebayes_screener(
  x,
  y,
  weights = NULL,
  resp.type,
  observations = NULL,
  logged2 = TRUE,
  method = "ls",
  design = NULL,
  contrasts = NULL,
  assay.type = c("array", "seq"),
  proportion = 0.01,
  coef = NULL,
  number = Inf,
  adjust.method = "BH",
  p.value = 1,
  top.table.sort.by = "none",
  return.type = c("pvalue", "object"),
  multi = c("combine", "max", "average", "sum", "raw"),
  logger = Logger(verbose = F),
  ...
)

Arguments

x

the input matrix, where rows are observations and columns are variables.

y

the response variable. Its number of rows must match the number of rows of x.

weights

priors of the observations

resp.type

the response type

observations

(optional) indices of observations to keep

logged2

whether data was logged 2.

method

fitting method; "ls" for least squares or "robust" for robust regression

design

the design matrix of the microarray experiment, with rows corresponding to samples and columns to coefficients to be estimated. Defaults to object$design if that is non-NULL, otherwise to the unit vector meaning that all samples will be treated as replicates of a single treatment group.

contrasts

numeric matrix with rows corresponding to coefficients in fit and columns containing contrasts. May be a vector if there is only one contrast. NAs are not allowed.

assay.type

Assay type: "array" for microarray data, "seq" for sequencing data

proportion

numeric value between 0 and 1, assumed proportion of genes which are differentially expressed

coef

(optional) an integer indicating the response variable to consider in multi-response data when multi = "raw"

number

maximum number of genes to list

adjust.method

method used to adjust the p-values for multiple testing. Options, in increasing conservatism, include "none", "BH", "BY" and "holm"

p.value

cutoff value for adjusted p-values. Only genes with lower p-values are listed.

top.table.sort.by

character string specifying statistic to rank genes by. Possible values for topTable are "logFC", "AveExpr", "t", "P", "p", "B" or "none". (Permitted synonyms are "M" for "logFC", "A" or "Amean" for "AveExpr", "T" for "t" and "p" for "P".)

multi

what to do when response has multiple output values

max

the max value of scores across multiple outputs is selected to get a single value for each observation

average

scores of multiple outputs are averaged to get a single value for each observation

sum

scores of multiple outputs are summed up to get a single value for each observation

raw

returns the scores for the multiple outputs

logger

a Logger

...

further arguments

Value

The p-value associated with each feature or NULL if the response type is not supported

a Screened object

See also

Author

Alessandro Barberis