Linear model for series of arrays

Fit linear model for each gene given a series of arrays

Usage

limma(
  x,
  design = NULL,
  method = "ls",
  weights = NULL,
  contrasts = NULL,
  assay.type = c("array", "seq"),
  proportion = 0.01,
  coef = NULL,
  number = Inf,
  adjust.method = "BH",
  p.value = 1,
  top.table.sort.by = "none",
  plotVenn = FALSE,
  outdir,
  logger = Logger()
)

Arguments

x

A matrix-like data object containing log-ratios or log-expression values for a series of arrays, with rows corresponding to genes and columns to samples.

design

the design matrix of the microarray experiment, with rows corresponding to arrays and columns to coefficients to be estimated. Defaults to the unit vector meaning that the arrays are treated as replicates.

method

fitting method; "ls" for least squares or "robust" for robust regression

contrasts

numeric matrix with rows corresponding to coefficients in fit and columns containing contrasts. May be a vector if there is only one contrast.

assay.type

Assay type: "array" for microarray data, "seq" for sequencing data

proportion

numeric value between 0 and 1, assumed proportion of genes which are differentially expressed

coef

column number or column name specifying which coefficient or contrast of the linear model is of interest. For topTable, can also be a vector of column subscripts, in which case the gene ranking is by F-statistic for that set of contrasts. If coef has a single value, then the moderated t-statistics and p-values for that coefficient or contrast are used. If coef takes two or more values, the moderated F-statistics for that set of coefficients or contrasts are used. If coef is left NULL, then all the coefficients or contrasts in the fitted model are used, except that any coefficient named (Intercept) will be removed.

number

maximum number of genes to list

adjust.method

method used to adjust the p-values for multiple testing. Options, in increasing conservatism, include "none", "BH", "BY" and "holm"

p.value

cutoff value for adjusted p-value

top.table.sort.by

character string specifying statistic to rank genes by. Possible values for topTable are "logFC", "AveExpr", "t", "P", "p", "B" or "none". (Permitted synonyms are "M" for "logFC", "A" or "Amean" for "AveExpr", "T" for "t" and "p" for "P".)

plotVenn

logical, whether to plot a Venn diagram showing numbers of genes signifcant in each comparison

outdir

output directory

logger

a Logger

warningsFile

A path to the warning file to use.

logFile

A path to the log file to use.

Value

A list with components (obtained from the different limma functions calls)