Supported Supervised Screening

This function returns a data.frame containing the currently supported supervised screening methods.

Usage

list_supported_supervised_screening_methods(resp.type, y)

Arguments

resp.type

(optional) the response type

y

(optional) the response variable

Value

A data.frame with 3 columns:

id

contains the id used in renoir for the supervised screening method (e.g. 'ebayes')

name

contains a short description of the method (e.g. 'empirical Bayes moderated t-statistics test')

Details

The currently implemented supervised screening methods are returned in a data.frame. If resp.type and y are provided, only the metrics supported for the given response are returned.

See also

Author

Alessandro Barberis