Plot PCA

Plot PCA

Usage

plot_PCA(
  PCA,
  groups,
  title = "",
  grid.arrange.ncol = 1,
  palette = NULL,
  palette.type = "qual",
  order = NULL,
  width = NULL,
  height = NULL,
  ...
)

Arguments

PCA

object as returned by prcomp

groups

data frame containing the groups (each column is a different group)

title

plot title

grid.arrange.ncol

number of columns to arrange plots in final figure

palette

(optional) when a single value, if a string, it is considered as a name of palette; if a number, it is the index into the lists of palettes of appropriate type. If palette is a single (unnamed) value, will be used for all the groups. If is a named vector, each element matching a column in groups will be used. The parameter is passed to scale_colour_brewer

palette.type

palette type. One of seq (sequential), div (diverging) or qual (qualitative)

order

the order to force before arranging the list of plots in the grid. If order is set to order = 'auto' and grid.arrange.ncol=2, the resulting grid has PC components from same group across the columns, while the same PC components for different groups are reported across the rows

width, height

size of each plot in inches

...

further arguments to arrangeGrob

Value

a ggplot2 object

Author

Alessandro Barberis