Significance analysis of microarrays (wrapper function to SAM)

Correlates a large number of features (eg genes) with an outcome variable, such as a group indicator, quantitative variable or survival time by using SAM function.

Usage

sam(
  x,
  y = NULL,
  censoring.status = NULL,
  resp.type = c("Quantitative", "Two class unpaired", "Survival", "Multiclass",
    "One class", "Two class paired", "Two class unpaired timecourse",
    "One class timecourse", "Two class paired timecourse", "Pattern discovery"),
  geneid = NULL,
  genenames = NULL,
  assay.type = c("array", "seq"),
  s0 = 1e-04,
  s0.perc = NULL,
  nperms = 100,
  center.arrays = FALSE,
  testStatistic = c("standard", "wilcoxon"),
  time.summary.type = c("slope", "signed.area"),
  regression.method = c("standard", "ranks"),
  return.x = FALSE,
  knn.neighbors = 10,
  random.seed = NULL,
  logged2 = FALSE,
  fdr.output = 0.2,
  eigengene.number = 1,
  nresamp = 20,
  plotSam = FALSE,
  verbose = FALSE,
  outdir = NULL,
  logger = Logger(),
  all.genes = FALSE
)

Arguments

x

Feature matrix: p (number of features) by n (number of samples), one observation per column (missing values allowed)

y

n-vector of outcome measurements

censoring.status

n-vector of censoring censoring.status (1=died or event occurred, 0=survived, or event was censored), needed for a censored survival outcome

resp.type

Problem type:

"Quantitative"

for a continuous parameter (Available for both array and sequencing data)

"Two class unpaired"

for both array and sequencing data

"Survival"

for censored survival outcome (for both array and sequencing data)

"Multiclass"

more than 2 groups (for both array and sequencing data)

"One class"

for a single group (only for array data)

"Two class paired"

for two classes with paired observations (for both array and sequencing data)

"Two class unpaired timecourse"

only for array data

"One class time course"

only for array data

"Two class.paired timecourse"

only for array data

"Pattern discovery"

only for array data

geneid

Optional character vector of geneids for output.

genenames

Optional character vector of genenames for output.

assay.type

Assay type: "array" for microarray data, "seq" for counts from sequencing

s0

Exchangeability factor for denominator of test statistic; Default in original function is a automatic choice. Only used for array data. Our default is a small number since sometimes the automatic choice could get weird results.

s0.perc

Percentile of standard deviation values to use for s0; default is automatic choice; -1 means s0=0 (different from s0.perc=0, meaning s0=zeroeth percentile of standard deviation values= min of sd values. Only used for array data.

nperms

Number of permutations used to estimate false discovery rates

center.arrays

Should the data for each sample (array) be median centered at the outset? Default =FALSE. Only used for array data.

testStatistic

Test statistic to use in two class unpaired case. Either "standard" (t-statistic) or ,"wilcoxon" (Two-sample wilcoxon or Mann-Whitney test). Only used for array data.

time.summary.type

Summary measure for each time course: "slope", or "signed.area"). Only used for array data.

return.x

= Should the matrix of feature values be returned? Only useful for time course data, where x contains summaries of the features over time. Otherwise x is the same as the input data data\$x

knn.neighbors

Number of nearest neighbors to use for imputation of missing features values. Only used for array data.

random.seed

Optional initial seed for random number generator (integer)

logged2

Has the data been transformed by log (base 2)? This information is used only for computing fold changes

fdr.output

(Approximate) False Discovery Rate cutoff for output in significant genes table

eigengene.number

Eigengene to be used (just for resp.type="Pattern discovery")

plotSam

whether to make Q-Q plot for SAM analysis, default TRUE.

outdir

output directory

logger

a Logger

all.genes

Should all genes be listed? Default FALSE

nresapm

For assay.type="seq", number of resamples used to construct test statistic. Default 20. Only used for sequencing data. Default is equal to nresamp and it must be at most nresamp. Only used for sequencing data

Value

A list with components (the same as returned by SAM)