Features screening

This function performs a feature screening, and return a Screened object

This function performs a feature screening by using the provided screeners. It combines the results from the different screeners and return a Screened object

Usage

screen(screener, ...)

# S4 method for Screener
screen(
  screener,
  x,
  y,
  weights = NULL,
  resp.type,
  observations = NULL,
  multi = c("max", "average", "sum"),
  order = c("increasing", "decreasing", "none"),
  cutoff = NULL,
  maxvars = NULL,
  logger,
  ...
)

# S4 method for ScreenerList
screen(
  screener,
  x,
  y,
  weights = NULL,
  resp.type,
  observations = NULL,
  combine = c("max", "average", "sum"),
  order = c("increasing", "decreasing"),
  cutoff = NULL,
  maxvars = NULL,
  logger = Logger(verbose = F),
  ...
)

Arguments

screener

a ScreenerList object

...

further arguments to screener function

x

the input matrix, where rows are observations and columns are variables.

y

the response variable. Its number of rows must match the number of rows of x.

weights

priors of the observations

resp.type

the response type

observations

(optional) indices of observations to keep

multi

what to do when response has multiple output values

max

the max value of scores across multiple outputs is selected to get a single value for each observation

average

scores of multiple outputs are averaged to get a single value for each observation

sum

scores of multiple outputs are summed up to get a single value for each observation

order

whether to order the results by the score values

increasing

results are sorted in increasing order

decreasing

results are sorted in decreasing order

none

results are not sorted

cutoff

(optional) numeric, the value to use as a threshold for filtering not significant variables

maxvars

integer, the maximum number of variables to keep. If the filtering is returning an higher number of variables, only the most significant maxvars variables are kept.

logger

a Logger

combine

how to combine the results from multiple screeners

max

the max value of scores across multiple Screened objects is selected to get a single value for each observation

average

scores of multiple Screened objects are averaged to get a single value for each observation

sum

scores of multiple Screened objects are summed up to get a single value for each observation

coef

(optional) an integer indicating the response variable to consider in multi-response data when multi = "raw"

Value

a Screened object

a Screened object

Author

Alessandro Barberis