Screener Class An S4 class providing the methods to perform a features screening.

This function performs a feature screening, and return the index of the retained elements in order of significance.

Usage

screening_default_gaussian(
  x,
  y,
  alternative = "two.sided",
  method = "pearson",
  conf.level = 0.95
)

Arguments

x

the input matrix, where rows are observations and columns are variables.

observations

(optional) indices of observations to keep

Value

the index of features to keep.

Details

The object consists of 10 slots

Slots

method

the type of screening to apply

default

if default is selected, the adopted test is different depending on the response type:

gaussian

test is correlation

mgaussian

test is correlation (for each gene, one correlation is computed with each response variable)

binomial

test is t-test or Wilcox test, depending on the number of observations

multinomial

test is anova or Kruskall-Wallis test, depending on the number of observations

poisson

test is correlation

cox

test is a Cox proportional hazards regression model

ebayes

permutation

resp.type

the response type

x

the input matrix, where rows are observations and columns are variables.

y

the response variable. Its number of rows must match the number of rows of x.

cutoff

numeric, the p-value to use as a threshold for filtering not significant variables.

coef

(optional) an integer indicating the response variable to consider in multi-response data

maxvars

integer, the maximum number of variables to keep. If the filtering is returning an higher number of variables, only the most significant maxvars variables are kept.

default

list, arguments to the function performing a default screening

ebayes

list, arguments to the function performing a screening based on empirical Bayes methods

permutation

list, arguments to the function performing a screening based on permutation

logger

a Logger

Author

Alessandro Barberis Features screening

Alessandro Barberis Compute the significance of the filtering test for the gaussian family