Data cols/rows filtering

This function filters the rows(/cols) of input data by removing the less variable ones. It uses the Interquartile Range (IQR) to measure variability.

variabilityFiltering.iqr(
  x,
  margin = 1,
  topThreshold = 0.25,
  verbose = TRUE,
  logFile = NULL,
  warningsFile = NULL
)

Arguments

x
A gene expression matrix where the rows represent the genes and the columns the samples.

margin
1 indicates row filtering, 2 indicates column filtering.

topThreshold
Select highly variable row(s)/col(s). Default is 0.25 (i.e. select row(s)/col(s) in top 25%).
0 returns the most variable col(s)/row(s), 1 the input matrix

verbose
logical. Should R report extra information on progress? Set to TRUE.

logFile
A path to the log file to use.

warningsFile
A path to the warning file to use.

Value

A list containing the processed matrix (x) and the number of removed rows (numRemovedRows) or columns (numRemovedCols).

Author

Alessandro Barberis